go to landing when back on first step
